random valid pool creation in each fuzz_target!
