+++ zrml/swaps/fuzz/exact_amount_pool_exiting.rs
@@ -3,30 +3,32 @@
 use libfuzzer_sys::fuzz_target;
 use zrml_swaps::mock::{ExtBuilder, Origin, Swaps};
 
-mod pool_creation;
-use pool_creation::{asset, get_valid_pool_id, ValidPoolData};
+mod data_structs;
+use data_structs::ExactAmountData;
+mod helper_functions;
+use helper_functions::asset;
+use zeitgeist_primitives::{traits::Swaps as SwapsTrait, types::ScoringRule};
 
-fuzz_target!(|data: ExactAmountPoolExitingData| {
+fuzz_target!(|data: ExactAmountData| {
     let mut ext = ExtBuilder::default().build();
     let _ = ext.execute_with(|| {
-        if let Ok(pool_id) = get_valid_pool_id(data.pool_creation) {
+        if let Ok(pool_id) = Swaps::create_pool(
+            data.pool_creation.origin.into(),
+            data.pool_creation.assets.into_iter().map(asset).collect(),
+            Some(data.pool_creation.base_asset).map(asset),
+            data.pool_creation.market_id.into(),
+            ScoringRule::CPMM,
+            Some(data.pool_creation.swap_fee).into(),
+            Some(data.pool_creation.weights).into(),
+        ) {
             let _ = Swaps::pool_exit_with_exact_pool_amount(
                 Origin::signed(data.origin.into()),
                 pool_id,
                 asset(data.asset),
                 data.pool_amount,
-                data.min_asset_amount,
+                data.asset_amount,
             );
         }
     });
     let _ = ext.commit_all();
 });
-
-#[derive(Debug, arbitrary::Arbitrary)]
-struct ExactAmountPoolExitingData {
-    origin: u8,
-    asset: (u128, u16),
-    pool_amount: u128,
-    min_asset_amount: u128,
-    pool_creation: ValidPoolData,
-}
