+++ zrml/swaps/fuzz/general_pool_exiting.rs
@@ -2,28 +2,32 @@
 
 use libfuzzer_sys::fuzz_target;
 use zrml_swaps::mock::{ExtBuilder, Origin, Swaps};
-mod pool_creation;
-use pool_creation::{get_valid_pool_id, ValidPoolData};
 
-fuzz_target!(|data: GeneralPoolExitingData| {
+mod data_structs;
+use data_structs::GeneralPoolData;
+use zeitgeist_primitives::{traits::Swaps as SwapsTrait, types::ScoringRule};
+mod helper_functions;
+use helper_functions::asset;
+
+fuzz_target!(|data: GeneralPoolData| {
     let mut ext = ExtBuilder::default().build();
     let _ = ext.execute_with(|| {
-        if let Ok(pool_id) = get_valid_pool_id(data.pool_creation) {
+        if let Ok(pool_id) = Swaps::create_pool(
+            data.pool_creation.origin.into(),
+            data.pool_creation.assets.into_iter().map(asset).collect(),
+            Some(data.pool_creation.base_asset).map(asset),
+            data.pool_creation.market_id.into(),
+            ScoringRule::CPMM,
+            Some(data.pool_creation.swap_fee).into(),
+            Some(data.pool_creation.weights).into(),
+        ) {
             let _ = Swaps::pool_exit(
                 Origin::signed(data.origin.into()),
                 pool_id,
                 data.pool_amount,
-                data.min_assets_out,
+                data.assets,
             );
         }
     });
     let _ = ext.commit_all();
 });
-
-#[derive(Debug, arbitrary::Arbitrary)]
-struct GeneralPoolExitingData {
-    origin: u8,
-    pool_amount: u128,
-    min_assets_out: Vec<u128>,
-    pool_creation: ValidPoolData,
-}
