+++ zrml/swaps/fuzz/general_pool_joining.rs
@@ -3,30 +3,33 @@
 use libfuzzer_sys::fuzz_target;
 
 use zrml_swaps::mock::{ExtBuilder, Origin, Swaps};
-mod pool_creation;
-use pool_creation::{get_valid_pool_id, ValidPoolData};
+mod data_structs;
+use data_structs::GeneralPoolData;
+use zeitgeist_primitives::{traits::Swaps as SwapsTrait, types::ScoringRule};
+mod helper_functions;
+use helper_functions::asset;
 
-fuzz_target!(|data: GeneralPoolJoiningData| {
+fuzz_target!(|data: GeneralPoolData| {
     let mut ext = ExtBuilder::default().build();
     let _ = ext.execute_with(|| {
-        if let Ok(pool_id) = get_valid_pool_id(data.pool_creation) {
+        if let Ok(pool_id) = Swaps::create_pool(
+            data.pool_creation.origin.into(),
+            data.pool_creation.assets.into_iter().map(asset).collect(),
+            Some(data.pool_creation.base_asset).map(asset),
+            data.pool_creation.market_id.into(),
+            ScoringRule::CPMM,
+            Some(data.pool_creation.swap_fee).into(),
+            Some(data.pool_creation.weights).into(),
+        ) {
             // join a pool with a valid pool id
             let _ = Swaps::pool_join(
                 Origin::signed(data.origin.into()),
                 pool_id,
                 data.pool_amount,
-                data.max_assets_in,
+                data.assets,
             );
         }
     });
 
     let _ = ext.commit_all();
 });
-
-#[derive(Debug, arbitrary::Arbitrary)]
-struct GeneralPoolJoiningData {
-    origin: u8,
-    pool_amount: u128,
-    max_assets_in: Vec<u128>,
-    pool_creation: ValidPoolData,
-}
