+++ zrml/swaps/fuzz/output_swap_exact_amount_pool_joining.rs
@@ -2,18 +2,30 @@
 
 use libfuzzer_sys::fuzz_target;
 use zrml_swaps::mock::{ExtBuilder, Origin, Swaps};
-mod pool_creation;
-use pool_creation::{asset, get_valid_pool_id, ValidPoolData};
 
-fuzz_target!(|data: OutputSwapExactAmountPoolJoiningData| {
+mod data_structs;
+use data_structs::SwapExactAmountData;
+mod helper_functions;
+use helper_functions::asset;
+use zeitgeist_primitives::{traits::Swaps as SwapsTrait, types::ScoringRule};
+
+fuzz_target!(|data: SwapExactAmountData| {
     let mut ext = ExtBuilder::default().build();
     let _ = ext.execute_with(|| {
-        if let Ok(pool_id) = get_valid_pool_id(data.pool_creation) {
+        if let Ok(pool_id) = Swaps::create_pool(
+            data.pool_creation.origin.into(),
+            data.pool_creation.assets.into_iter().map(asset).collect(),
+            Some(data.pool_creation.base_asset).map(asset),
+            data.pool_creation.market_id.into(),
+            ScoringRule::CPMM,
+            Some(data.pool_creation.swap_fee).into(),
+            Some(data.pool_creation.weights).into(),
+        ) {
             let _ = Swaps::swap_exact_amount_out(
                 Origin::signed(data.origin.into()),
                 pool_id,
                 asset(data.asset_in),
-                data.max_amount_asset_in,
+                data.asset_amount_in,
                 asset(data.asset_out),
                 data.asset_amount_out,
                 data.max_price,
@@ -22,14 +34,3 @@ fuzz_target!(|data: OutputSwapExactAmountPoolJoiningData| {
     });
     let _ = ext.commit_all();
 });
-
-#[derive(Debug, arbitrary::Arbitrary)]
-struct OutputSwapExactAmountPoolJoiningData {
-    origin: u8,
-    asset_in: (u128, u16),
-    max_amount_asset_in: u128,
-    asset_out: (u128, u16),
-    asset_amount_out: u128,
-    max_price: u128,
-    pool_creation: ValidPoolData,
-}
